requirements.txt
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
dj-database-url==0.5.0
Django==2.1.5
django-cors-headers==2.5.2
djangorestframework==3.9.1
gunicorn==19.9.0
httpie==1.0.2
idna==2.8
Pillow==5.2.0
psycopg2==2.8.2
Pygments==2.3.1
python-decouple==3.1
pytz==2019.1
requests==2.21.0
urllib3==1.24.2
whitenoise==4.1.2
